Rename helpers.js to utils.js It's consistent with general naming pattern in Gutenberg.
